{"types":"C","runs":1}

+++ Problem21.java
@@ -3,11 +3,11 @@ import acm.program.GraphicsProgram;
 
 public class Problem21 extends GraphicsProgram{
 	public void run(){
-		for(int m = 0; m < getWidth(); m = m + getWidth()/10){
+		for(double m = 0; m < getWidth(); m = m + getWidth()/10){
 			GLine line = new GLine(m, 0, m, getHeight());
 			add(line);
 		}
-		for(int n = 0; n < getHeight(); n = n + getHeight()/10){
+		for(double n = 0; n < getHeight(); n = n + getHeight()/10){
 			GLine HorLine = new GLine(0, n, getWidth(), n);
 			add(HorLine);
 		}
